Feature/register/fe

See merge request s04-final/s04p31a404!39

Former-commit-id: c452879c922e120fcf45df1aecc6e5e87833d139

+++ backend/src/main/java/com/ssafy/eurekka/repository/BarcodeRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface BarcodeRepository extends MongoRepository<Barcode, ObjectId> {
 
-  Optional<Barcode> findByCode(int code);
+  Optional<Barcode> findByCode(Double code);
 }
